+++ HISTORY.rst
@@ -13,6 +13,7 @@ v0.11.0 (2018-09-14)
 * Fix bug with ANOVA implementation being cached with dropped doses
 * Save history of models executed when dropping doses
 * Update reporting (xlsx) to include prior sessions when doses are dropped
+* Update reporting (docx) to better document sessions where doses are dropped
 
 v0.10.0 (2018-04-11)
 --------------------